- more refactoring  (according to specs):   * more change from docker to dock   * change sortAttributes to sortBy - implemented getDocks

+++ ScheduleShipManagementServiceConsumer/src/main/kotlin/com/example/scheduleshipmanagementserviceconsumer/Size.kt
@@ -1,8 +1,8 @@
 package com.example.scheduleshipmanagementserviceconsumer
 
 class Size() {
-    var width: Int? = null
-    var length: Int? = null
+    var width: Float? = null
+    var length: Float? = null
 
     override fun toString(): String {
         return "SizeBoundary(width=$width, length=$length)"
